apps/worker/requirements.txt
redis==5.0.8
boto3==1.35.20
psycopg2-binary==2.9.9
